mips: fix after moving gic/gcr files
CONFIG_MIPS_GIC is added to default-configs/mips{64}{el}-softmmu.mak
