+++ default-configs/mips64el-softmmu.mak
@@ -39,3 +39,4 @@ CONFIG_MC146818RTC=y
 CONFIG_VT82C686=y
 CONFIG_ISA_TESTDEV=y
 CONFIG_EMPTY_SLOT=y
+CONFIG_MIPS_GIC=y
